requirements.txt
paramiko
requests
pyyaml